Refactor app_router.dart: Update pageBuilder functions and add imports
Refactor the app_router.dart file to update the pageBuilder functions for the '/' and '/buttons' routes. The pageBuilder functions now include the necessary imports for the HomeScreen and ButtonsScreen. Additionally, the pageBuilder functions now pass the corresponding screen widgets as children to the _animationConfig function. This change ensures that the correct screens are displayed when navigating to the '/' and '/buttons' routes.

Refactor buttons_screnn.dart: Add name constant

In the buttons_screnn.dart file, add a name constant for the ButtonsScreen class. This constant can be used to refer to the screen by its name in other parts of the code.

Refactor cards_screen.dart: Add name constant

In the cards_screen.dart file, add a name constant for the CardsScreen class. This constant can be used to refer to the screen by its name in other parts of the code.

Refactor home_screen.dart: Add name constant and remove unnecessary imports

In the home_screen.dart file, add a name constant for the HomeScreen class. This constant can be used to refer to the screen by its name in other parts of the code. Additionally, remove unnecessary imports for the app_router.dart and buttons_screnn.dart files.

Refactor main.dart: Update app theme and home screen

Refactor the main.dart file to update the app theme and home screen. This change ensures that the app has an updated theme and the correct home screen is displayed.